MicroBlaze UART & I²C EEPROM Interface
This is an ongoing FPGA-based  project that integrates a MicroBlaze soft processor with UART and I²C peripherals to communicate with an external EEPROM memory device.
The hardware platform is built in Vivado 2022.1, and the software application is under active development in Vitis to handle EEPROM read/write operations and serial terminal communication.